Add `opac-error` command line tool
Located in opacplot.scripts.opac_error, opac-error will compare two EoS/Opacity
tables and plot their RMS/Absolute percent error. It also has logging and
plotting capabilities. This was done to help locate potential errors/bugs
in the conversion tools. It will also be useful for anyone with multiple
EoS/Opacity tables for the same material who wants to see the differences
between the two.

As of now, it does not have documentation. In the future, documentation needs
to be added, both API and on the README for GitHub.