Add sesame-extract clt; update README; update setup.py
sesame-extract is a command line tool that extracts individual
materials from SESAME tables based on material ID. opac-error
and opac-convert have unexpected behavior when opening
SESAME files with very many materials, and so sesame-extract
addresses this issue by creating files with only a single
material. Information regarding this issue and some sesame-extract
usage examples were added to the README. A link to the script for
the command line tools was added to setup.py.